% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contr.dummy.R
\name{contr.ordinal}
\alias{contr.ordinal}
\title{Internal Re-export of contr.ordinal}
\description{
This function is re-exported from the \code{kknn} package to ensure compatibility with the \code{train.kknn} function,
which requires \code{contr.ordinal} to be available in the current namespace. This function generates dummy variables
for ordered factors.
}
\keyword{internal}
