% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.vswift.R
\name{print}
\alias{print}
\alias{print.vswift}
\title{Print Parameter Information and/or Model Evaluation Metrics}
\usage{
\method{print}{vswift}(x, configs = TRUE, metrics = TRUE, models = NULL, ...)
}
\arguments{
\item{x}{A vswift object.}

\item{configs}{A logical value indicating whether to print model configuration information from the vswift
object. Default is \code{TRUE}.}

\item{metrics}{A logical value indicating whether to print model evaluation metrics from the vswift object. If
\code{TRUE}, precision, recall, and F1 scores for each class will be displayed, along with their mean values
(if cross-validation was used). Default is \code{TRUE}.}

\item{models}{A character string or a character vector specifying the classification algorithm(s) information to be
printed. If \code{NULL}, all model information will be printed. The following options are available:
\itemize{
 \item \code{"lda"}: Linear Discriminant Analysis
 \item \code{"qda"}: Quadratic Discriminant Analysis
 \item \code{"logistic"}: Unregularized Logistic Regression
 \item \code{"regularized_logistic"}: Regularized Logistic Regression
 \item \code{"svm"}: Support Vector Machine
 \item \code{"naivebayes"}: Naive Bayes
 \item \code{"nnet"}: Neural Network
 \item \code{"knn"}: K-Nearest Neighbors
 \item \code{"decisiontree"}: Decision Tree
 \item \code{"randomforest"}: Random Forest
 \item \code{"multinom"}: Unregularized Multinomial Logistic Regression
 \item \code{"regularized_multinomial"}: Regularized Multinomial Logistic Regression
 \item \code{"xgboost"}: Extreme Gradient Boosting
 }
 Default = \code{NULL}.}

\item{...}{No additional arguments are currently supported.}
}
\description{
Prints model configuration details and/or model evaluation metrics (classification accuracy, precision,
recall, and F1 scores) from a vswift object.
}
\examples{
# Load an example dataset

data(iris)

# Perform a train-test split with an 80\% training set using LDA

result <- classCV(
  data = iris,
  target = "Species",
  models = "lda",
  train_params = list(split = 0.8, stratified = TRUE, random_seed = 50)
)

#  Print parameter information and performance metrics
print(result)

}
\author{
Donisha Smith
}
