% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_vswift.R
\name{print.vswift}
\alias{print.vswift}
\title{print.vswift}
\usage{
\method{print}{vswift}(object, parameters = FALSE, metrics = FALSE)
}
\arguments{
\item{object}{An object of class vswift.}

\item{parameters}{A logical value to print parameter information from vswift object. Default = FALSE.}

\item{metrics}{A logical value to print model evaluation metrics from vswift object. Default = FALSE.}
}
\description{
print.vswift is print parameter information and/or model evaluation metrics from a vswift object.
}
