% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_split.R
\name{stratified_split}
\alias{stratified_split}
\title{stratified_split}
\usage{
stratified_split(
  data = NULL,
  y_col = NULL,
  split = NULL,
  fold_n = NULL,
  stratified = FALSE,
  random_seed = NULL,
  create_data = FALSE
)
}
\arguments{
\item{data}{A data frame.}

\item{y_col}{A numerical index or character name for the response variable.}

\item{split}{A numerical value between 0.5 to 0.9 indicating the proportion of data to use for the training set, leaving the rest for the test set. If not specified, train-test splitting will not be done.}

\item{fold_n}{A numerical value between 3-30 indicating the number of k-folds. If left empty, k-fold cross validation will not be performed.}

\item{stratified}{A logical value indicating if stratified sampling should be used. Default = FALSE.}

\item{random_seed}{A numerical value for the random seed to be used. Default = NULL.}

\item{create_data}{A logical value indicating whether to create all training and test/validation data frames. Default = FALSE.}
}
\value{
A list containing the indices for train-test splitting and/or k-fold cross-validation, with information on the class distribution in the training, test sets, and folds (if applicable)
        as well as the generated split datasets and folds based on the indices.
}
\description{
Perform train-test splitting or k-fold cross-validation on data using stratified sampling.
}
\examples{
# Load example dataset 

data(iris)

# Obtain indices for 80\% training/test split and 5-fold CV

stratified_list <- stratified_split(data = iris, y_col = "Species", split = 0.8, fold_n = 5)
}
