% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vswift.R
\name{plot.vswift}
\alias{plot.vswift}
\title{plot.vswift}
\usage{
\method{plot}{vswift}(object, split = TRUE, cv = TRUE)
}
\arguments{
\item{object}{An object of class vswift.}

\item{split}{A logical value indicating whether to plot metrics for train-test splitting results. Default = TRUE.}

\item{cv}{A logical value indicating whether to plot metrics for k-fold cross-validation results. Note: Solid blueline represents the mean
and dashed red line represents the standard deviation. Default = TRUE.}
}
\value{
Plots representing evaluation metrics.
}
\description{
Plot model evaluation metrics from a vswift object.
}
\examples{
# Load an example dataset

data(iris)

# Perform a train-test split with an 80\% training set and stratified_sampling using QDA

result <- categorical_cv_split(data = iris, y_col = "Species", split = 0.8,
model_type = "qda", stratified = TRUE)

# Plot performance metrics for train-test split

plot(result)

}
