% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vswift.R
\name{plot.vswift}
\alias{plot.vswift}
\title{plot.vswift}
\usage{
\method{plot}{vswift}(object, split = FALSE, cv = FALSE)
}
\arguments{
\item{object}{An object of class vswift.}

\item{split}{A logical value to plot metrics for train-test splitting results. Default = FALSE.}

\item{cv}{A logical value to plot metrics for k-fold cross validation results. Default = FALSE.}
}
\description{
plot.vswift plot model evaluation metrics from a vswift object.
}
