% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_vswift.R
\name{print.vswift}
\alias{print.vswift}
\title{Print parameter information and/or model evaluation metrics}
\usage{
\method{print}{vswift}(object, parameters = TRUE, metrics = TRUE)
}
\arguments{
\item{object}{An object of class vswift.}

\item{parameters}{A logical value indicating whether to print parameter information from the vswift object. Default = TRUE.}

\item{metrics}{A logical value indicating whether to print model evaluation metrics from the vswift object. This will display the precision, recall, and f-score for each class.
If the vswift object contains information for train-test splitting, the classification accuracy for the training and test set as well as the precision, recall, and f-score for each class
will be displayed. If the vswift object contains information for k-fold validation, the mean and standard error for the classification accuracy and class' precision, recall, and f-score will be displayed. Default = TRUE.}
}
\description{
`print.vswift` prints parameter information and/or model evaluation metrics (classification accuracy and precision, recall, and f-score for each class) from a vswift object.
}
\examples{
# Load an example dataset

data(iris)

# Perform a train-test split with an 80\% training set using LDA

result <- categorical_cv_split(data = iris, target = "Species", split = 0.8,
model_type = "lda", stratified = TRUE, random_seed = 123)

#  Print parameter information and performance metrics
print(result)

}
