% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_split.R
\name{stratified_split}
\alias{stratified_split}
\title{Create split datasets and/or folds with optional stratification}
\usage{
stratified_split(
  data = NULL,
  target = NULL,
  split = NULL,
  n_folds = NULL,
  stratified = FALSE,
  random_seed = NULL,
  create_data = FALSE
)
}
\arguments{
\item{data}{A data frame.}

\item{target}{A numerical index or character name for the target variable.}

\item{split}{A numerical value between 0.5 to 0.9 indicating the proportion of data to use for the training set, leaving the rest for the test set. If not specified, train-test splitting will not be done.}

\item{n_folds}{A numerical value between 3-30 indicating the number of k-folds. If left empty, k-fold cross validation will not be performed.}

\item{stratified}{A logical value indicating if stratified sampling should be used. Default = FALSE.}

\item{random_seed}{A numerical value for the random seed to be used. Default = NULL.}

\item{create_data}{A logical value indicating whether to create all training and test/validation data frames. Default = FALSE.}
}
\value{
A list containing the indices for train-test splitting and/or k-fold cross-validation, with information on the class distribution in the training, test sets, and folds (if applicable)
        as well as the generated split datasets and folds based on the indices.
}
\description{
`stratified_split` generates train-test split datasets and/or k-fold cross-validation folds, with the option to perform stratified sampling based on class distribution.
}
\examples{
# Load example dataset 

data(iris)

# Obtain indices for 80\% training/test split and 5-fold CV

stratified_list <- stratified_split(data = iris, target = "Species", split = 0.8, n_folds = 5)
}
