% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\docType{package}
\name{vswift}
\alias{vswift}
\title{vswift}
\description{
A package that allows for model evaluation using train-test splitting and k-fold cross validation using several popular classification
 algorithms such as logistic regression, linear discriminant analysis, quadratic discriminant analysis, support vector machines, k-nearest neighbors, 
 artificial neural network, decision trees, and random forest.
}
\details{
vswift

A package for simple classification model evaluation.
}
\author{
Donisha Smith
}
